Move MakeWithValue.hpp to from DataStructures to Utilities
Fix #472
