+++ src/DataStructures/Tensor/EagerMath/DeterminantAndInverse.hpp
@@ -10,9 +10,9 @@
 #include <utility>
 
 #include "DataStructures/DataVector.hpp"
-#include "DataStructures/MakeWithValue.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "ErrorHandling/Assert.hpp"
+#include "Utilities/MakeWithValue.hpp"
 #include "Utilities/Requires.hpp"
 
 namespace determinant_and_inverse_detail {
