+++ src/Domain/CoordinateMaps/Identity.cpp
@@ -3,10 +3,10 @@
 
 #include "Domain/CoordinateMaps/Identity.hpp"
 
-#include "DataStructures/MakeWithValue.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "ErrorHandling/Assert.hpp"
 #include "Utilities/DereferenceWrapper.hpp"
+#include "Utilities/MakeWithValue.hpp"
 
 namespace CoordinateMaps {
