+++ src/Domain/CoordinateMaps/ProductMaps.hpp
@@ -8,12 +8,12 @@
 
 #include <memory>
 
-#include "DataStructures/MakeWithValue.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "DataStructures/Tensor/TypeAliases.hpp"
 #include "Parallel/CharmPupable.hpp"
 #include "Parallel/PupStlCpp11.hpp"
 #include "Utilities/DereferenceWrapper.hpp"
+#include "Utilities/MakeWithValue.hpp"
 
 namespace CoordinateMaps {
