+++ src/PointwiseFunctions/AnalyticSolutions/EinsteinSolutions/KerrSchild.cpp
@@ -8,11 +8,11 @@
 #include <utility>
 
 #include "DataStructures/DataVector.hpp"
-#include "DataStructures/MakeWithValue.hpp"
 #include "DataStructures/Tensor/EagerMath/Magnitude.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "PointwiseFunctions/GeneralRelativity/GrTags.hpp"
 #include "Utilities/Gsl.hpp"
+#include "Utilities/MakeWithValue.hpp"
 
 namespace EinsteinSolutions {
