+++ src/PointwiseFunctions/AnalyticSolutions/EinsteinSolutions/Minkowski.hpp
@@ -3,9 +3,9 @@
 
 #pragma once
 
-#include "DataStructures/MakeWithValue.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "Options/Options.hpp"
+#include "Utilities/MakeWithValue.hpp"
 
 namespace EinsteinSolutions {
