+++ src/PointwiseFunctions/MathFunctions/PowX.cpp
@@ -6,8 +6,8 @@
 #include <cmath>
 
 #include "DataStructures/DataVector.hpp"
-#include "DataStructures/MakeWithValue.hpp"
 #include "Utilities/ConstantExpressions.hpp"
+#include "Utilities/MakeWithValue.hpp"
 
 namespace MathFunctions {
